requirements.txt
requests
beautifulsoup4
selenium
Pillow
lxml
discord
httpx
msgpack